creacion de componente TestimonialCard, Press usa TestimonialCard
